feat(desktop): split settings into app/workspace/session surfaces
App settings = full-page studio (app-only, no rail); workspace + session settings
become one-page divide-y panes (instant save, no footer). Three legacy dialogs
deleted; entry points rewired to event dispatch. ProviderChip-backed override pick.
