+++ apps/desktop/src/features/settings/components/SettingsStudio/index.tsx
@@ -1,0 +1,24 @@
+import { Settings } from 'lucide-react';
+import { StudioShell } from '../../../../shared/components/StudioShell';
+import { AppScopePanel } from './AppScopePanel';
+
+type Props = {
+  readonly initialFocus?: string;
+  readonly onClose: () => void;
+};
+
+export const SettingsStudio = ({ initialFocus, onClose }: Props) => (
+  <StudioShell
+    icon={Settings}
+    title="Settings"
+    workspaceName="App settings"
+    closeLabel="close settings"
+    onClose={onClose}
+  >
+    {(requestClose) => (
+      <div className="min-h-0 flex-1">
+        <AppScopePanel initialSection={initialFocus} requestClose={requestClose} />
+      </div>
+    )}
+  </StudioShell>
+);
